market and training forms done
