+++ app/api/trainings/route.js
@@ -1,3 +1,4 @@
+import db from "@/lib/db";
 import { NextResponse } from "next/server";
 
 export async function POST(request) {
@@ -11,21 +12,56 @@ export async function POST(request) {
       isActive,
       content,
     } = await request.json();
-    const newTraining = {
-      title,
-      slug,
-      categoryId,
-      imageUrl,
-      description,
-      isActive,
-      content,
-    };
-    console.log(newTraining);
+    const existingTraining = await db.training.findUnique({
+      where: {
+        slug,
+      },
+    });
+    if (existingTraining) {
+      return NextResponse.json(
+        {
+          data: null,
+          message: "Training already exists",
+        },
+        { status: 409 }
+      );
+    }
+    const newTraining = await db.training.create({
+      data: {
+        title,
+        slug,
+        categoryId,
+        imageUrl,
+        description,
+        isActive,
+        content,
+      },
+    });
     return NextResponse.json(newTraining);
   } catch (error) {
     console.log(error);
     return NextResponse.json(
       { maessage: "failed to create training", error },
+      { status: 500 }
+    );
+  }
+}
+
+export async function GET(request) {
+  try {
+    const trainings = await db.training
+      .findMany
+      //   {
+      //   orderBy: {
+      //     createdAt: "desc",
+      //   },
+      // }
+      ();
+    return NextResponse.json(trainings);
+  } catch (error) {
+    console.log(error);
+    return NextResponse.json(
+      { maessage: "failed to fetch training", error },
       { status: 500 }
     );
   }
