+++ prisma/schema.prisma
@@ -93,6 +93,8 @@ model Training {
   content     String
   categoryId  String     @db.ObjectId
   category    Category   @relation(fields: [categoryId], references: [id])
+    createdAt           DateTime   @default(now())
+  updatedAt           DateTime?  @updatedAt
 }
 
 enum UserRole {
